+++ .gitignore
@@ -3,3 +3,12 @@ games_training_data/
 *.pgn
 *.7z
 *.bz2
+__pycache__/
+static/pngs/b 2.png
+static/pngs/N.png
+static/pngs/P.png
+static/pngs/B.png
+static/pngs/R.png
+static/pngs/k.png
+static/pngs/q.png
+
