+++ docs/PERFORMANCE_OPTIMIZATION.md
