+++ experiments/UCI_engine_async.py
@@ -1,15 +1,16 @@
 #!/usr/bin/env python3
 """
-UCI Protocol Implementation for AlphaZero Chess Engine - CUDA/Windows Optimized
+UCI Protocol Implementation for AlphaZero Chess Engine with Async Neural Network
 
-High-performance UCI engine optimized for Windows systems with NVIDIA GPUs.
-Specifically tuned for RTX 4080 and similar high-end consumer GPUs.
+This module implements the Universal Chess Interface (UCI) protocol for the AlphaZero
+chess engine with asynchronous neural network evaluation for much higher performance.
 
-Time management: Dynamically adjusts rollouts based on available time,
-accounting for the much higher throughput of CUDA-optimized implementation.
+Time management: Dynamically adjusts the number of MCTS rollouts based on available
+time, using measured performance from the async engine.
 """
 
 import sys
+sys.path.append('..')
 import os
 import chess
 import torch
@@ -18,21 +19,22 @@ import time
 import threading
 from queue import Queue
 import AlphaZeroNetwork
-from MCTS_ultra_performance import UltraPerformanceMCTSEngine
+from MCTS.MCTS_async import MCTSEngine
+from async_neural_net_server import NeuralNetworkPool
 from device_utils import get_optimal_device, optimize_for_device
 
 
-class TimeManagerCUDA:
-    """Time management optimized for high-performance CUDA systems."""
+class TimeManager:
+    """Manages time allocation for moves based on game time constraints."""
     
-    def __init__(self, base_rollouts=50000, base_time=1.0, threads=64):
+    def __init__(self, base_rollouts=2000, base_time=1.0, threads=32):
         """
-        Initialize time manager for CUDA systems.
+        Initialize time manager.
         
         Args:
-            base_rollouts: Expected rollouts in base_time (50k for RTX 4080)
+            base_rollouts: Number of rollouts that take base_time seconds
             base_time: Time in seconds for base_rollouts
-            threads: Number of worker threads
+            threads: Number of threads available
         """
         self.base_rollouts = base_rollouts
         self.base_time = base_time
@@ -45,7 +47,7 @@ class TimeManagerCUDA:
         
     def update_performance(self, rollouts, elapsed_time):
         """Update measured performance based on actual timing."""
-        if elapsed_time > 0.05:  # Only update for meaningful measurements
+        if elapsed_time > 0.1:  # Only update for meaningful measurements
             new_rps = rollouts / elapsed_time
             if self.measured_rollouts_per_second is None:
                 self.measured_rollouts_per_second = new_rps
@@ -57,7 +59,7 @@ class TimeManagerCUDA:
             
     def get_rollouts_per_second(self):
         """Get the best estimate of rollouts per second."""
-        if self.measured_rollouts_per_second and self.measurement_count >= 2:
+        if self.measured_rollouts_per_second and self.measurement_count >= 3:
             return self.measured_rollouts_per_second
         return self.rollouts_per_second
         
@@ -81,8 +83,8 @@ class TimeManagerCUDA:
         increment = winc if turn else binc
         
         if time_left is None:
-            # No time limit, use high quality search
-            return min(100000, self.base_rollouts * 2)
+            # No time limit, use default
+            return self.base_rollouts
             
         # Convert to seconds
         time_left_sec = time_left / 1000.0
@@ -91,95 +93,90 @@ class TimeManagerCUDA:
         # Calculate time to allocate for this move
         if movestogo and movestogo > 0:
             # Time control with moves to go
-            time_per_move = time_left_sec / movestogo + increment_sec * 0.9
+            time_per_move = time_left_sec / movestogo + increment_sec * 0.8
         else:
             # Sudden death or unknown moves to go
-            if time_left_sec > 300:  # 5+ minutes
-                time_fraction = 0.025  # Use 2.5% when lots of time
-            elif time_left_sec > 60:
-                time_fraction = 0.04   # Use 4% with moderate time
+            # Use a fraction of remaining time, scaling down as time decreases
+            if time_left_sec > 60:
+                time_fraction = 0.04  # Use 4% when we have plenty of time
             elif time_left_sec > 10:
-                time_fraction = 0.08   # Use 8% when time is lower
+                time_fraction = 0.06  # Use 6% when time is getting lower
             else:
-                time_fraction = 0.15   # Use 15% when very low on time
+                time_fraction = 0.10  # Use 10% when very low on time
                 
-            time_per_move = time_left_sec * time_fraction + increment_sec * 0.9
+            time_per_move = time_left_sec * time_fraction + increment_sec * 0.8
         
-        # Reserve time for overhead (less needed with CUDA)
-        time_per_move = time_per_move * 0.95 - 0.05
+        # Add safety buffer for overhead (communication, model loading, etc)
+        # Reserve at least 100ms for overhead
+        time_per_move = time_per_move * 0.9 - 0.1
         
         # Ensure minimum thinking time
         time_per_move = max(0.05, time_per_move)
         
-        # Cap at 30% of remaining time
-        time_per_move = min(time_per_move, time_left_sec * 0.3)
+        # Ensure we don't use more than 40% of remaining time
+        time_per_move = min(time_per_move, time_left_sec * 0.4)
         
-        # Calculate rollouts based on performance
+        # Calculate rollouts based on available time
         rps = self.get_rollouts_per_second()
         rollouts = int(rps * time_per_move)
         
-        # Quality thresholds for CUDA systems
-        rollouts = max(5000, rollouts)    # Minimum 5k for quality
-        rollouts = min(500000, rollouts)  # Cap at 500k
+        # Ensure minimum rollouts for quality
+        rollouts = max(1000, rollouts)  # Higher minimum for async engine
+        
+        # Cap maximum rollouts to prevent excessive thinking
+        rollouts = min(1000000, rollouts)  # Higher cap for async engine
         
         return rollouts
 
 
-class UCIEngineCUDA:
-    """High-performance UCI engine for Windows/CUDA systems."""
+class UCIEngineAsync:
+    """UCI Protocol handler for AlphaZero chess engine with async neural network."""
     
-    def __init__(self, model_path=None, threads=64, verbose=False, 
-                 max_batch_size=512, device_id=0):
+    def __init__(self, model_path=None, threads=32, verbose=False, max_batch_size=256):
         """
-        Initialize UCI engine for CUDA.
+        Initialize UCI engine.
         
         Args:
             model_path: Path to the neural network model file
-            threads: Number of threads for MCTS (64 optimal for high-end CPUs)
+            threads: Number of threads to use for MCTS
             verbose: Whether to output debug information
-            max_batch_size: Maximum batch size (512 optimal for RTX 4080)
-            device_id: CUDA device ID to use
+            max_batch_size: Maximum batch size for neural network
         """
         self.model_path = model_path
         self.threads = threads
         self.verbose = verbose
         self.max_batch_size = max_batch_size
-        self.device_id = device_id
         self.board = chess.Board()
         self.model = None
         self.mcts_engine = None
+        self.nn_pool = None
         self.device = None
-        self.time_manager = TimeManagerCUDA(base_rollouts=50000, threads=threads)
+        self.time_manager = TimeManager(base_rollouts=5000, threads=threads)
         self.search_thread = None
         self.stop_search = threading.Event()
         self.best_move = None
-        self.move_overhead = 20  # Lower overhead for CUDA
+        self.move_overhead = 30  # Default move overhead in milliseconds
         
-        # Windows-specific optimizations
-        if sys.platform == 'win32':
-            # Set process priority to high
-            try:
-                import psutil
-                p = psutil.Process()
-                p.nice(psutil.HIGH_PRIORITY_CLASS)
-            except:
-                pass
-                
     def load_model(self):
-        """Load the neural network model and initialize CUDA engine."""
+        """Load the neural network model and initialize MCTS engine."""
         try:
             if not self.model_path:
+                # Use default model if none specified
                 self.model_path = "AlphaZeroNet_20x256_distributed.pt"
             
-            # Handle path resolution for Windows
+            # Try to find the model file
             if not os.path.isabs(self.model_path):
+                # Try relative to current directory first
                 if os.path.exists(self.model_path):
-                    full_path = os.path.abspath(self.model_path)
+                    full_path = self.model_path
                 else:
+                    # Try relative to script directory
                     script_dir = os.path.dirname(os.path.abspath(__file__))
                     full_path = os.path.join(script_dir, self.model_path)
                     if not os.path.exists(full_path):
                         print(f"info string ERROR: Model file not found: {self.model_path}")
+                        print(f"info string Tried: {os.path.abspath(self.model_path)}")
+                        print(f"info string Tried: {full_path}")
                         sys.stdout.flush()
                         return False
             else:
@@ -189,47 +186,43 @@ class UCIEngineCUDA:
                     sys.stdout.flush()
                     return False
                 
-            # Set CUDA device
-            if torch.cuda.is_available():
-                if self.device_id < torch.cuda.device_count():
-                    self.device = torch.device(f'cuda:{self.device_id}')
-                    device_name = torch.cuda.get_device_name(self.device_id)
-                    device_memory = torch.cuda.get_device_properties(self.device_id).total_memory / 1024**3
-                else:
-                    print(f"info string ERROR: CUDA device {self.device_id} not available")
-                    sys.stdout.flush()
-                    return False
-            else:
-                print("info string ERROR: CUDA not available. This version requires an NVIDIA GPU.")
-                sys.stdout.flush()
-                return False
-                
+            self.device, device_str = get_optimal_device()
             if self.verbose:
                 print(f"info string Loading model from: {full_path}")
-                print(f"info string Using CUDA device {self.device_id}: {device_name}")
-                print(f"info string GPU Memory: {device_memory:.1f}GB")
+                print(f"info string Loading model on device: {device_str}")
                 sys.stdout.flush()
                 
-            # Load model
             self.model = AlphaZeroNetwork.AlphaZeroNet(20, 256)
             weights = torch.load(full_path, map_location=self.device)
             self.model.load_state_dict(weights)
+            self.model = optimize_for_device(self.model, self.device)
             self.model.eval()
             
-            # Initialize Ultra-Performance MCTS engine
-            self.mcts_engine = UltraPerformanceMCTSEngine(
+            # Disable gradients for inference
+            for param in self.model.parameters():
+                param.requires_grad = False
+                
+            # Initialize Neural Network Pool
+            self.nn_pool = NeuralNetworkPool(
                 self.model,
-                device=self.device,
-                batch_size=self.max_batch_size,
-                num_workers=self.threads,
+                self.device,
+                num_workers=min(4, self.threads // 8),  # Fewer NN workers
+                batch_size=self.max_batch_size
+            )
+            self.nn_pool.start()
+            
+            # Initialize MCTS engine
+            self.mcts_engine = MCTSEngine(
+                self.nn_pool,
+                num_parallel_games=1,
+                threads_per_game=self.threads,
                 verbose=self.verbose
             )
-            self.mcts_engine.start()
                 
             if self.verbose:
                 print(f"info string Model loaded successfully")
-                print(f"info string CUDA MCTS engine initialized")
-                print(f"info string Workers: {self.threads}, Batch size: {self.max_batch_size}")
+                print(f"info string Async MCTS engine initialized with {self.threads} workers")
+                print(f"info string Max batch size: {self.max_batch_size}")
                 sys.stdout.flush()
             return True
             
@@ -240,14 +233,13 @@ class UCIEngineCUDA:
             
     def uci(self):
         """Handle 'uci' command."""
-        print("id name AlphaZero UCI Engine (CUDA/Windows)")
+        print("id name AlphaZero UCI Engine (Async)")
         print("id author AlphaZero Bot")
-        print("option name Threads type spin default 64 min 1 max 256")
+        print("option name Threads type spin default 32 min 1 max 256")
         print("option name Model type string default AlphaZeroNet_20x256_distributed.pt")
         print("option name Verbose type check default false")
-        print("option name Move Overhead type spin default 20 min 0 max 5000")
-        print("option name Max Batch Size type spin default 512 min 32 max 2048")
-        print("option name Device ID type spin default 0 min 0 max 7")
+        print("option name Move Overhead type spin default 30 min 0 max 5000")
+        print("option name Max Batch Size type spin default 256 min 8 max 1024")
         print("uciok")
         sys.stdout.flush()
         
@@ -255,12 +247,18 @@ class UCIEngineCUDA:
         """Handle 'isready' command."""
         if self.model is None or self.mcts_engine is None:
             if not self.load_model():
+                # Model loading failed, but we still need to respond
                 pass
         print("readyok")
         sys.stdout.flush()
         
     def position(self, args):
-        """Handle 'position' command."""
+        """
+        Handle 'position' command.
+        
+        Args:
+            args: List of position arguments
+        """
         if len(args) < 1:
             return
             
@@ -288,11 +286,17 @@ class UCIEngineCUDA:
                         print(f"info string Invalid move: {move_str}")
                         
     def search_position(self, rollouts):
-        """Search current position using CUDA-optimized MCTS."""
+        """
+        Search current position using async MCTS.
+        
+        Args:
+            rollouts: Number of rollouts to perform
+        """
         self.stop_search.clear()
         self.best_move = None
         
         try:
+            # Check if we have a model
             if self.mcts_engine is None:
                 print(f"info string ERROR: No MCTS engine loaded")
                 sys.stdout.flush()
@@ -300,19 +304,15 @@ class UCIEngineCUDA:
                 
             start_time = time.time()
             
-            # Periodic progress updates
-            update_interval = max(1000, rollouts // 20)
-            last_update = 0
-            
             # Run search
-            best_move = self.mcts_engine.search(
+            best_move, root = self.mcts_engine.search(
                 self.board, 
                 rollouts
             )
             
             elapsed_time = time.time() - start_time
             
-            # Update time manager
+            # Update time manager with actual performance
             if elapsed_time > 0:
                 self.time_manager.update_performance(rollouts, elapsed_time)
                 
@@ -331,12 +331,13 @@ class UCIEngineCUDA:
                 # Output final info
                 nps = int(rollouts / elapsed_time) if elapsed_time > 0 else 0
                 print(f"info depth {rollouts} nodes {rollouts} nps {nps} pv {self.best_move}")
-                        sys.stdout.flush()
-                        
+                sys.stdout.flush()
+                
                 if self.verbose:
                     print(f"info string Completed {rollouts} rollouts in {elapsed_time:.2f}s")
-                    print(f"info string Nodes per second: {rollouts/elapsed_time:.0f}")
-                    
+                    print(f"info string Nodes per second: {rollouts/elapsed_time:.1f}")
+                    sys.stdout.flush()
+                        
                 # Clean up
                 root.cleanup()
                 
@@ -347,7 +348,12 @@ class UCIEngineCUDA:
             sys.stdout.flush()
                 
     def go(self, args):
-        """Handle 'go' command."""
+        """
+        Handle 'go' command.
+        
+        Args:
+            args: List of go arguments
+        """
         # Parse time control parameters
         wtime = None
         btime = None
@@ -383,15 +389,18 @@ class UCIEngineCUDA:
             else:
                 i += 1
                 
-        # Calculate rollouts
+        # Calculate rollouts based on time
         if infinite:
-            rollouts = 1000000  # 1M for analysis
+            rollouts = 1000000  # High number for analysis
         elif movetime:
+            # Fixed time per move
             time_sec = movetime / 1000.0
+            # Account for move overhead
             time_sec = max(0.1, time_sec - self.move_overhead / 1000.0)
-            rollouts = int(self.time_manager.get_rollouts_per_second() * time_sec * 0.95)
+            rollouts = int(self.time_manager.rollouts_per_second * time_sec * 0.95)
         else:
-            # Adjust for overhead
+            # Calculate based on game time
+            # Adjust time for move overhead
             if wtime is not None:
                 wtime = max(100, wtime - self.move_overhead)
             if btime is not None:
@@ -416,10 +425,23 @@ class UCIEngineCUDA:
         if self.best_move:
             print(f"bestmove {self.best_move}")
         else:
-            # Fallback
+            # Fallback: pick first legal move that doesn't cause threefold repetition
             legal_moves = list(self.board.legal_moves)
-            if legal_moves:
-                print(f"bestmove {legal_moves[0]}")
+            fallback_move = None
+            
+            for move in legal_moves:
+                test_board = self.board.copy()
+                test_board.push(move)
+                if not test_board.can_claim_threefold_repetition():
+                    fallback_move = move
+                    break
+            
+            # If all moves lead to threefold repetition, use the first legal move
+            if fallback_move is None and legal_moves:
+                fallback_move = legal_moves[0]
+                
+            if fallback_move:
+                print(f"bestmove {fallback_move}")
         sys.stdout.flush()
                 
     def stop(self):
@@ -434,16 +456,21 @@ class UCIEngineCUDA:
     def quit(self):
         """Handle 'quit' command."""
         self.stop_search.set()
-        if self.mcts_engine:
-            self.mcts_engine.stop()
+        if self.nn_pool:
+            self.nn_pool.stop()
         sys.exit(0)
         
     def setoption(self, args):
-        """Handle 'setoption' command."""
+        """
+        Handle 'setoption' command.
+        
+        Args:
+            args: List of option arguments
+        """
         if len(args) < 4 or args[0] != "name":
             return
             
-        # Find value position
+        # Find where "value" appears in args
         value_idx = -1
         for i, arg in enumerate(args):
             if arg == "value":
@@ -453,6 +480,7 @@ class UCIEngineCUDA:
         if value_idx == -1 or value_idx < 2:
             return
             
+        # Reconstruct name and value allowing for multi-word names
         name = " ".join(args[1:value_idx]).lower()
         value = " ".join(args[value_idx + 1:])
         
@@ -460,6 +488,7 @@ class UCIEngineCUDA:
             try:
                 self.threads = int(value)
                 self.time_manager.threads = self.threads
+                # Recreate MCTS engine with new thread count
                 if self.mcts_engine:
                     self.mcts_engine.stop()
                     self.mcts_engine = None
@@ -467,7 +496,7 @@ class UCIEngineCUDA:
                 pass
         elif name == "model":
             self.model_path = value
-            self.model = None
+            self.model = None  # Force reload on next isready
             if self.mcts_engine:
                 self.mcts_engine.stop()
                 self.mcts_engine = None
@@ -481,14 +510,7 @@ class UCIEngineCUDA:
         elif name == "max batch size":
             try:
                 self.max_batch_size = int(value)
-                if self.mcts_engine:
-                    self.mcts_engine.stop()
-                    self.mcts_engine = None
-            except:
-                pass
-        elif name == "device id":
-            try:
-                self.device_id = int(value)
+                # Recreate MCTS engine with new batch size
                 if self.mcts_engine:
                     self.mcts_engine.stop()
                     self.mcts_engine = None
@@ -497,11 +519,6 @@ class UCIEngineCUDA:
             
     def run(self):
         """Main UCI protocol loop."""
-        # Windows line buffering
-        if sys.platform == 'win32':
-            import msvcrt
-            msvcrt.setmode(sys.stdout.fileno(), os.O_BINARY)
-            
         while True:
             try:
                 line = input().strip()
@@ -526,6 +543,7 @@ class UCIEngineCUDA:
                 elif command == "setoption":
                     self.setoption(parts[1:])
                 elif command == "ucinewgame":
+                    # Reset board for new game
                     self.board = chess.Board()
                 elif self.verbose:
                     print(f"info string Unknown command: {command}")
@@ -542,28 +560,29 @@ class UCIEngineCUDA:
 
 def main():
     """Main entry point."""
+    # Ensure stdout is line-buffered for proper UCI communication
+    if hasattr(sys.stdout, 'reconfigure'):
+        sys.stdout.reconfigure(line_buffering=True)
+    
     parser = argparse.ArgumentParser(
-        description="High-performance UCI engine for Windows/CUDA systems"
+        description="UCI Protocol wrapper for AlphaZero chess engine with async neural network"
     )
     parser.add_argument("--model", help="Path to model file", 
                        default="AlphaZeroNet_20x256_distributed.pt")
     parser.add_argument("--threads", type=int, help="Number of threads", 
-                       default=64)
+                       default=32)
     parser.add_argument("--verbose", action="store_true", 
                        help="Enable verbose output")
-    parser.add_argument("--batch-size", type=int, help="Max batch size",
-                       default=512)
-    parser.add_argument("--device", type=int, help="CUDA device ID",
-                       default=0)
+    parser.add_argument("--batch-size", type=int, help="Max batch size for neural network",
+                       default=256)
     
     args = parser.parse_args()
     
-    engine = UCIEngineCUDA(
+    engine = UCIEngineAsync(
         model_path=args.model,
         threads=args.threads,
         verbose=args.verbose,
-        max_batch_size=args.batch_size,
-        device_id=args.device
+        max_batch_size=args.batch_size
     )
     
     try:
