+++ experiments/UCI_engine_cuda.py
@@ -1,15 +1,16 @@
 #!/usr/bin/env python3
 """
-UCI Protocol Implementation for AlphaZero Chess Engine with Async Neural Network
+UCI Protocol Implementation for AlphaZero Chess Engine - CUDA/Windows Optimized
 
-This module implements the Universal Chess Interface (UCI) protocol for the AlphaZero
-chess engine with asynchronous neural network evaluation for much higher performance.
+High-performance UCI engine optimized for Windows systems with NVIDIA GPUs.
+Specifically tuned for RTX 4080 and similar high-end consumer GPUs.
 
-Time management: Dynamically adjusts the number of MCTS rollouts based on available
-time, using measured performance from the async engine.
+Time management: Dynamically adjusts rollouts based on available time,
+accounting for the much higher throughput of CUDA-optimized implementation.
 """
 
 import sys
+sys.path.append('..')
 import os
 import chess
 import torch
@@ -18,21 +19,22 @@ import time
 import threading
 from queue import Queue
 import AlphaZeroNetwork
-from MCTS_ultra_performance import UltraPerformanceMCTSEngine
+from MCTS.MCTS_cuda import MCTSEngineCUDA
+from async_neural_net_server_cuda import NeuralNetworkPoolCUDA
 from device_utils import get_optimal_device, optimize_for_device
 
 
-class TimeManager:
-    """Manages time allocation for moves based on game time constraints."""
+class TimeManagerCUDA:
+    """Time management optimized for high-performance CUDA systems."""
     
-    def __init__(self, base_rollouts=2000, base_time=1.0, threads=32):
+    def __init__(self, base_rollouts=50000, base_time=1.0, threads=64):
         """
-        Initialize time manager.
+        Initialize time manager for CUDA systems.
         
         Args:
-            base_rollouts: Number of rollouts that take base_time seconds
+            base_rollouts: Expected rollouts in base_time (50k for RTX 4080)
             base_time: Time in seconds for base_rollouts
-            threads: Number of threads available
+            threads: Number of worker threads
         """
         self.base_rollouts = base_rollouts
         self.base_time = base_time
@@ -45,7 +47,7 @@ class TimeManager:
         
     def update_performance(self, rollouts, elapsed_time):
         """Update measured performance based on actual timing."""
-        if elapsed_time > 0.1:  # Only update for meaningful measurements
+        if elapsed_time > 0.05:  # Only update for meaningful measurements
             new_rps = rollouts / elapsed_time
             if self.measured_rollouts_per_second is None:
                 self.measured_rollouts_per_second = new_rps
@@ -57,7 +59,7 @@ class TimeManager:
             
     def get_rollouts_per_second(self):
         """Get the best estimate of rollouts per second."""
-        if self.measured_rollouts_per_second and self.measurement_count >= 3:
+        if self.measured_rollouts_per_second and self.measurement_count >= 2:
             return self.measured_rollouts_per_second
         return self.rollouts_per_second
         
@@ -81,8 +83,8 @@ class TimeManager:
         increment = winc if turn else binc
         
         if time_left is None:
-            # No time limit, use default
-            return self.base_rollouts
+            # No time limit, use high quality search
+            return min(100000, self.base_rollouts * 2)
             
         # Convert to seconds
         time_left_sec = time_left / 1000.0
@@ -91,89 +93,96 @@ class TimeManager:
         # Calculate time to allocate for this move
         if movestogo and movestogo > 0:
             # Time control with moves to go
-            time_per_move = time_left_sec / movestogo + increment_sec * 0.8
+            time_per_move = time_left_sec / movestogo + increment_sec * 0.9
         else:
             # Sudden death or unknown moves to go
-            # Use a fraction of remaining time, scaling down as time decreases
-            if time_left_sec > 60:
-                time_fraction = 0.04  # Use 4% when we have plenty of time
+            if time_left_sec > 300:  # 5+ minutes
+                time_fraction = 0.025  # Use 2.5% when lots of time
+            elif time_left_sec > 60:
+                time_fraction = 0.04   # Use 4% with moderate time
             elif time_left_sec > 10:
-                time_fraction = 0.06  # Use 6% when time is getting lower
+                time_fraction = 0.08   # Use 8% when time is lower
             else:
-                time_fraction = 0.10  # Use 10% when very low on time
+                time_fraction = 0.15   # Use 15% when very low on time
                 
-            time_per_move = time_left_sec * time_fraction + increment_sec * 0.8
+            time_per_move = time_left_sec * time_fraction + increment_sec * 0.9
         
-        # Add safety buffer for overhead (communication, model loading, etc)
-        # Reserve at least 100ms for overhead
-        time_per_move = time_per_move * 0.9 - 0.1
+        # Reserve time for overhead (less needed with CUDA)
+        time_per_move = time_per_move * 0.95 - 0.05
         
         # Ensure minimum thinking time
         time_per_move = max(0.05, time_per_move)
         
-        # Ensure we don't use more than 40% of remaining time
-        time_per_move = min(time_per_move, time_left_sec * 0.4)
+        # Cap at 30% of remaining time
+        time_per_move = min(time_per_move, time_left_sec * 0.3)
         
-        # Calculate rollouts based on available time
+        # Calculate rollouts based on performance
         rps = self.get_rollouts_per_second()
         rollouts = int(rps * time_per_move)
         
-        # Ensure minimum rollouts for quality
-        rollouts = max(1000, rollouts)  # Higher minimum for async engine
-        
-        # Cap maximum rollouts to prevent excessive thinking
-        rollouts = min(1000000, rollouts)  # Higher cap for async engine
+        # Quality thresholds for CUDA systems
+        rollouts = max(5000, rollouts)    # Minimum 5k for quality
+        rollouts = min(500000, rollouts)  # Cap at 500k
         
         return rollouts
 
 
-class UCIEngineAsync:
-    """UCI Protocol handler for AlphaZero chess engine with async neural network."""
+class UCIEngineCUDA:
+    """High-performance UCI engine for Windows/CUDA systems."""
     
-    def __init__(self, model_path=None, threads=32, verbose=False, max_batch_size=256):
+    def __init__(self, model_path=None, threads=64, verbose=False, 
+                 max_batch_size=512, device_id=0):
         """
-        Initialize UCI engine.
+        Initialize UCI engine for CUDA.
         
         Args:
             model_path: Path to the neural network model file
-            threads: Number of threads to use for MCTS
+            threads: Number of threads for MCTS (64 optimal for high-end CPUs)
             verbose: Whether to output debug information
-            max_batch_size: Maximum batch size for neural network
+            max_batch_size: Maximum batch size (512 optimal for RTX 4080)
+            device_id: CUDA device ID to use
         """
         self.model_path = model_path
         self.threads = threads
         self.verbose = verbose
         self.max_batch_size = max_batch_size
+        self.device_id = device_id
         self.board = chess.Board()
         self.model = None
         self.mcts_engine = None
+        self.nn_pool = None
         self.device = None
-        self.time_manager = TimeManager(base_rollouts=5000, threads=threads)
+        self.time_manager = TimeManagerCUDA(base_rollouts=50000, threads=threads)
         self.search_thread = None
         self.stop_search = threading.Event()
         self.best_move = None
-        self.move_overhead = 30  # Default move overhead in milliseconds
+        self.move_overhead = 20  # Lower overhead for CUDA
         
+        # Windows-specific optimizations
+        if sys.platform == 'win32':
+            # Set process priority to high
+            try:
+                import psutil
+                p = psutil.Process()
+                p.nice(psutil.HIGH_PRIORITY_CLASS)
+            except:
+                pass
+                
     def load_model(self):
-        """Load the neural network model and initialize MCTS engine."""
+        """Load the neural network model and initialize CUDA engine."""
         try:
             if not self.model_path:
-                # Use default model if none specified
                 self.model_path = "AlphaZeroNet_20x256_distributed.pt"
             
-            # Try to find the model file
+            # Handle path resolution for Windows
             if not os.path.isabs(self.model_path):
-                # Try relative to current directory first
                 if os.path.exists(self.model_path):
-                    full_path = self.model_path
+                    full_path = os.path.abspath(self.model_path)
                 else:
-                    # Try relative to script directory
                     script_dir = os.path.dirname(os.path.abspath(__file__))
                     full_path = os.path.join(script_dir, self.model_path)
                     if not os.path.exists(full_path):
                         print(f"info string ERROR: Model file not found: {self.model_path}")
-                        print(f"info string Tried: {os.path.abspath(self.model_path)}")
-                        print(f"info string Tried: {full_path}")
                         sys.stdout.flush()
                         return False
             else:
@@ -183,36 +192,54 @@ class UCIEngineAsync:
                     sys.stdout.flush()
                     return False
                 
-            self.device, device_str = get_optimal_device()
+            # Set CUDA device
+            if torch.cuda.is_available():
+                if self.device_id < torch.cuda.device_count():
+                    self.device = torch.device(f'cuda:{self.device_id}')
+                    device_name = torch.cuda.get_device_name(self.device_id)
+                    device_memory = torch.cuda.get_device_properties(self.device_id).total_memory / 1024**3
+                else:
+                    print(f"info string ERROR: CUDA device {self.device_id} not available")
+                    sys.stdout.flush()
+                    return False
+            else:
+                print("info string ERROR: CUDA not available. This version requires an NVIDIA GPU.")
+                sys.stdout.flush()
+                return False
+                
             if self.verbose:
                 print(f"info string Loading model from: {full_path}")
-                print(f"info string Loading model on device: {device_str}")
+                print(f"info string Using CUDA device {self.device_id}: {device_name}")
+                print(f"info string GPU Memory: {device_memory:.1f}GB")
                 sys.stdout.flush()
                 
+            # Load model
             self.model = AlphaZeroNetwork.AlphaZeroNet(20, 256)
             weights = torch.load(full_path, map_location=self.device)
             self.model.load_state_dict(weights)
-            self.model = optimize_for_device(self.model, self.device)
             self.model.eval()
             
-            # Disable gradients for inference
-            for param in self.model.parameters():
-                param.requires_grad = False
-                
-            # Initialize Ultra-Performance MCTS engine
-            self.mcts_engine = UltraPerformanceMCTSEngine(
+            # Initialize Neural Network Pool
+            self.nn_pool = NeuralNetworkPoolCUDA(
                 self.model,
-                device=self.device,
-                batch_size=self.max_batch_size,
-                num_workers=self.threads,
+                self.device,
+                num_workers=min(4, self.threads // 16),  # Fewer NN workers for CUDA
+                batch_size=self.max_batch_size
+            )
+            self.nn_pool.start()
+            
+            # Initialize MCTS engine
+            self.mcts_engine = MCTSEngineCUDA(
+                self.nn_pool,
+                num_parallel_games=1,
+                threads_per_game=self.threads,
                 verbose=self.verbose
             )
-            self.mcts_engine.start()
                 
             if self.verbose:
                 print(f"info string Model loaded successfully")
-                print(f"info string Async MCTS engine initialized with {self.threads} workers")
-                print(f"info string Max batch size: {self.max_batch_size}")
+                print(f"info string CUDA MCTS engine initialized")
+                print(f"info string Workers: {self.threads}, Batch size: {self.max_batch_size}")
                 sys.stdout.flush()
             return True
             
@@ -223,13 +250,14 @@ class UCIEngineAsync:
             
     def uci(self):
         """Handle 'uci' command."""
-        print("id name AlphaZero UCI Engine (Async)")
+        print("id name AlphaZero UCI Engine (CUDA/Windows)")
         print("id author AlphaZero Bot")
-        print("option name Threads type spin default 32 min 1 max 256")
+        print("option name Threads type spin default 64 min 1 max 256")
         print("option name Model type string default AlphaZeroNet_20x256_distributed.pt")
         print("option name Verbose type check default false")
-        print("option name Move Overhead type spin default 30 min 0 max 5000")
-        print("option name Max Batch Size type spin default 256 min 8 max 1024")
+        print("option name Move Overhead type spin default 20 min 0 max 5000")
+        print("option name Max Batch Size type spin default 512 min 32 max 2048")
+        print("option name Device ID type spin default 0 min 0 max 7")
         print("uciok")
         sys.stdout.flush()
         
@@ -237,18 +265,12 @@ class UCIEngineAsync:
         """Handle 'isready' command."""
         if self.model is None or self.mcts_engine is None:
             if not self.load_model():
-                # Model loading failed, but we still need to respond
                 pass
         print("readyok")
         sys.stdout.flush()
         
     def position(self, args):
-        """
-        Handle 'position' command.
-        
-        Args:
-            args: List of position arguments
-        """
+        """Handle 'position' command."""
         if len(args) < 1:
             return
             
@@ -276,17 +298,11 @@ class UCIEngineAsync:
                         print(f"info string Invalid move: {move_str}")
                         
     def search_position(self, rollouts):
-        """
-        Search current position using async MCTS.
-        
-        Args:
-            rollouts: Number of rollouts to perform
-        """
+        """Search current position using CUDA-optimized MCTS."""
         self.stop_search.clear()
         self.best_move = None
         
         try:
-            # Check if we have a model
             if self.mcts_engine is None:
                 print(f"info string ERROR: No MCTS engine loaded")
                 sys.stdout.flush()
@@ -294,16 +310,19 @@ class UCIEngineAsync:
                 
             start_time = time.time()
             
-            # Get root node for progressive output
-            root = self.mcts_engine.search(
+            # Periodic progress updates
+            update_interval = max(1000, rollouts // 20)
+            last_update = 0
+            
+            # Run search
+            best_move, root = self.mcts_engine.search(
                 self.board, 
-                rollouts, 
-                return_root=True
+                rollouts
             )
             
             elapsed_time = time.time() - start_time
             
-            # Update time manager with actual performance
+            # Update time manager
             if elapsed_time > 0:
                 self.time_manager.update_performance(rollouts, elapsed_time)
                 
@@ -322,13 +341,12 @@ class UCIEngineAsync:
                 # Output final info
                 nps = int(rollouts / elapsed_time) if elapsed_time > 0 else 0
                 print(f"info depth {rollouts} nodes {rollouts} nps {nps} pv {self.best_move}")
-                sys.stdout.flush()
-                
+                        sys.stdout.flush()
+                        
                 if self.verbose:
                     print(f"info string Completed {rollouts} rollouts in {elapsed_time:.2f}s")
-                    print(f"info string Nodes per second: {rollouts/elapsed_time:.1f}")
-                    sys.stdout.flush()
-                        
+                    print(f"info string Nodes per second: {rollouts/elapsed_time:.0f}")
+                    
                 # Clean up
                 root.cleanup()
                 
@@ -339,12 +357,7 @@ class UCIEngineAsync:
             sys.stdout.flush()
                 
     def go(self, args):
-        """
-        Handle 'go' command.
-        
-        Args:
-            args: List of go arguments
-        """
+        """Handle 'go' command."""
         # Parse time control parameters
         wtime = None
         btime = None
@@ -380,18 +393,15 @@ class UCIEngineAsync:
             else:
                 i += 1
                 
-        # Calculate rollouts based on time
+        # Calculate rollouts
         if infinite:
-            rollouts = 1000000  # High number for analysis
+            rollouts = 1000000  # 1M for analysis
         elif movetime:
-            # Fixed time per move
             time_sec = movetime / 1000.0
-            # Account for move overhead
             time_sec = max(0.1, time_sec - self.move_overhead / 1000.0)
-            rollouts = int(self.time_manager.rollouts_per_second * time_sec * 0.95)
+            rollouts = int(self.time_manager.get_rollouts_per_second() * time_sec * 0.95)
         else:
-            # Calculate based on game time
-            # Adjust time for move overhead
+            # Adjust for overhead
             if wtime is not None:
                 wtime = max(100, wtime - self.move_overhead)
             if btime is not None:
@@ -416,23 +426,10 @@ class UCIEngineAsync:
         if self.best_move:
             print(f"bestmove {self.best_move}")
         else:
-            # Fallback: pick first legal move that doesn't cause threefold repetition
+            # Fallback
             legal_moves = list(self.board.legal_moves)
-            fallback_move = None
-            
-            for move in legal_moves:
-                test_board = self.board.copy()
-                test_board.push(move)
-                if not test_board.can_claim_threefold_repetition():
-                    fallback_move = move
-                    break
-            
-            # If all moves lead to threefold repetition, use the first legal move
-            if fallback_move is None and legal_moves:
-                fallback_move = legal_moves[0]
-                
-            if fallback_move:
-                print(f"bestmove {fallback_move}")
+            if legal_moves:
+                print(f"bestmove {legal_moves[0]}")
         sys.stdout.flush()
                 
     def stop(self):
@@ -447,21 +444,16 @@ class UCIEngineAsync:
     def quit(self):
         """Handle 'quit' command."""
         self.stop_search.set()
-        if self.mcts_engine:
-            self.mcts_engine.stop()
+        if self.nn_pool:
+            self.nn_pool.stop()
         sys.exit(0)
         
     def setoption(self, args):
-        """
-        Handle 'setoption' command.
-        
-        Args:
-            args: List of option arguments
-        """
+        """Handle 'setoption' command."""
         if len(args) < 4 or args[0] != "name":
             return
             
-        # Find where "value" appears in args
+        # Find value position
         value_idx = -1
         for i, arg in enumerate(args):
             if arg == "value":
@@ -471,7 +463,6 @@ class UCIEngineAsync:
         if value_idx == -1 or value_idx < 2:
             return
             
-        # Reconstruct name and value allowing for multi-word names
         name = " ".join(args[1:value_idx]).lower()
         value = " ".join(args[value_idx + 1:])
         
@@ -479,17 +470,18 @@ class UCIEngineAsync:
             try:
                 self.threads = int(value)
                 self.time_manager.threads = self.threads
-                # Recreate MCTS engine with new thread count
-                if self.mcts_engine:
-                    self.mcts_engine.stop()
+                if self.nn_pool:
+                    self.nn_pool.stop()
+                    self.nn_pool = None
                     self.mcts_engine = None
             except:
                 pass
         elif name == "model":
             self.model_path = value
-            self.model = None  # Force reload on next isready
-            if self.mcts_engine:
-                self.mcts_engine.stop()
+            self.model = None
+            if self.nn_pool:
+                self.nn_pool.stop()
+                self.nn_pool = None
                 self.mcts_engine = None
         elif name == "verbose":
             self.verbose = value.lower() in ["true", "yes", "1"]
@@ -501,15 +493,29 @@ class UCIEngineAsync:
         elif name == "max batch size":
             try:
                 self.max_batch_size = int(value)
-                # Recreate MCTS engine with new batch size
-                if self.mcts_engine:
-                    self.mcts_engine.stop()
+                if self.nn_pool:
+                    self.nn_pool.stop()
+                    self.nn_pool = None
+                    self.mcts_engine = None
+            except:
+                pass
+        elif name == "device id":
+            try:
+                self.device_id = int(value)
+                if self.nn_pool:
+                    self.nn_pool.stop()
+                    self.nn_pool = None
                     self.mcts_engine = None
             except:
                 pass
             
     def run(self):
         """Main UCI protocol loop."""
+        # Windows line buffering
+        if sys.platform == 'win32':
+            import msvcrt
+            msvcrt.setmode(sys.stdout.fileno(), os.O_BINARY)
+            
         while True:
             try:
                 line = input().strip()
@@ -534,7 +540,6 @@ class UCIEngineAsync:
                 elif command == "setoption":
                     self.setoption(parts[1:])
                 elif command == "ucinewgame":
-                    # Reset board for new game
                     self.board = chess.Board()
                 elif self.verbose:
                     print(f"info string Unknown command: {command}")
@@ -551,29 +556,28 @@ class UCIEngineAsync:
 
 def main():
     """Main entry point."""
-    # Ensure stdout is line-buffered for proper UCI communication
-    if hasattr(sys.stdout, 'reconfigure'):
-        sys.stdout.reconfigure(line_buffering=True)
-    
     parser = argparse.ArgumentParser(
-        description="UCI Protocol wrapper for AlphaZero chess engine with async neural network"
+        description="High-performance UCI engine for Windows/CUDA systems"
     )
     parser.add_argument("--model", help="Path to model file", 
                        default="AlphaZeroNet_20x256_distributed.pt")
     parser.add_argument("--threads", type=int, help="Number of threads", 
-                       default=32)
+                       default=64)
     parser.add_argument("--verbose", action="store_true", 
                        help="Enable verbose output")
-    parser.add_argument("--batch-size", type=int, help="Max batch size for neural network",
-                       default=256)
+    parser.add_argument("--batch-size", type=int, help="Max batch size",
+                       default=512)
+    parser.add_argument("--device", type=int, help="CUDA device ID",
+                       default=0)
     
     args = parser.parse_args()
     
-    engine = UCIEngineAsync(
+    engine = UCIEngineCUDA(
         model_path=args.model,
         threads=args.threads,
         verbose=args.verbose,
-        max_batch_size=args.batch_size
+        max_batch_size=args.batch_size,
+        device_id=args.device
     )
     
     try:
