+++ experiments/playchess_multigpu.py
@@ -1,3 +1,5 @@
+import sys
+sys.path.append('..')
 import argparse
 import chess
 import torch
