+++ test/test_mcts_comparison.py
@@ -7,6 +7,8 @@ the Ultra-Performance version maintains similar playing strength while
 achieving higher NPS (nodes per second).
 """
 
+import sys
+sys.path.append('..')
 import chess
 import torch
 import argparse
@@ -15,7 +17,7 @@ from concurrent.futures import ThreadPoolExecutor
 import numpy as np
 from AlphaZeroNetwork import AlphaZeroNet
 import MCTS
-from MCTS_ultra_performance import UltraPerformanceMCTSEngine
+from experiments.MCTS.MCTS_ultra_performance import UltraPerformanceMCTSEngine
 import device_utils
 
 class MCTSComparison:
